* commit '450bef3b9c2c0cd4136aa1f60e3b8ec6f49f3419': (22 commits)
  OR-2413 Run a dummy DAV server for calendar midway tests.
  OR-2413 Make test-modules-frontend pass.
  OR-2413 Make test-frontend pass.
  OR-2413 Make test-modules-midway pass.
  OR-2413 Fixed unit-test following express 4 res.send() deprecation.
  OR-2413 Make test-modules-unit-backend pass.
  OR-2413 Make test-midway-backend pass.
  OR-2413 Make test-unit-storage pass.
  OR-2413 Make test-unit-backend pass
  OR-2413 Removed jscs, it has merged with eslint.
  OR-2413 Async v2 has breaking changes in filter, some and every. https://github.com/caolan/async/blob/master/CHANGELOG.md#breaking-changes
  OR-2413 Express 4 deprecates res.send(status).
  OR-2413 Express 4 deprecates res.json(status, object)
  OR-2413 Express 4 has deprecated req.param()
  OR-2413 Offset and limits need to be numbers in MongoDB 3.2. See https://jira.mongodb.org/browse/DOCS-6995
  OR-2413 Accomodated changed for newer morgan versions.
  OR-2413 Accomodated changes to newer body-parser versions. https://github.com/expressjs/body-parser#api
  OR-2413 Accomodated changes for newer versions of express-session. https://github.com/expressjs/session#sessionoptions
  OR-2413 Do not use old Schema.Mixed of mongoose.
  OR-2413 Accomodate jsonwebtoken callback change. https://github.com/auth0/node-jsonwebtoken/pull/192
  ...

+++ test/midway-backend/all.js
@@ -27,7 +27,9 @@ before(function() {
       fs.writeFileSync(tmpPath + '/db.json', JSON.stringify({connectionString: 'mongodb://' + host + ':' + testConfig.mongodb.port + '/' + testConfig.mongodb.dbname, connectionOptions: {auto_reconnect: false}}));
     },
     removeDBConfigFile: function() {
-      fs.unlinkSync(tmpPath + '/db.json');
+      if (fs.existsSync(tmpPath + '/db.json')) {
+        fs.unlinkSync(tmpPath + '/db.json');
+      }
     },
     initCore: function(callback) {
       var core = require(basePath + '/backend/core');
@@ -40,6 +42,8 @@ before(function() {
     },
     initRedisConfiguration: function(mongoose, callback) {
       var configuration = require('../../backend/core/esn-config');
+
+      mongoose.Promise = require('q').Promise; // http://mongoosejs.com/docs/promises.html
       mongoose.connect(this.mongoUrl);
       var self = this;
 
@@ -75,6 +79,11 @@ after(function() {
   }
   delete process.env.NODE_CONFIG;
   delete process.env.NODE_ENV;
+});
+
+// https://github.com/mfncooper/mockery/issues/34
+before(function() {
+  require('canvas');
 });
 
 beforeEach(function() {
